Update REVLib vendordep

+++ src/main/java/frc/robot/Constants.java
@@ -347,10 +347,9 @@ public final class Constants {
     public static final int BOTTOM_ROLLER_PWM_PORT = 1;
     public static final int DOOR_CLOSED_LIMIT_DIO_PORT = 0;
 
-    public static final int INTAKE_SPEED = -1;
+    public static final double INTAKE_SPEED = -1;
     public static final double INTAKE_DOOR_ANGLE = 40.0;
-    public static final double DOOR_KP =
-        0.01; // Making this big so it go weeeee, the last kp was 0.01
+    public static final double DOOR_KP = 0.01;
     public static final double DOOR_KI = 0;
     public static final double DOOR_KD = 0;
     public static final double DOOR_ENCODERS_TO_DEGREES = 360 / 90;
